utils/requirements.txt
# 記事自動作成システム 必要パッケージ

anthropic==0.40.0
requests==2.31.0
Pillow==10.4.0
python-dotenv==1.0.0
